requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Marshmallow==1.2.1
Flask-RESTx==1.3.0
Flask-SQLAlchemy==3.1.1
jwt==1.3.1
psycopg2==2.9.9
PyJWT==2.9.0
Werkzeug==3.0.3
SQLAlchemy==2.0.32
Flask-Migrate==4.0.7
